refactor: move card
