+++ packages/components/card/src/card.visualroute.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { css } from '@emotion/core';
 import { Card } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../test/percy';
+import { Suite, Spec } from '../../../../test/percy';
 
 export const routePath = '/card';
